ser: Do not transform strings to include single quotes
This change drops the special logic in serialization to add single quotes to
certain strings passed into `serialize_str`. Let's leave this up to the user to
determine if they want to add special quotes. This gets rid of the problem I was
seeing with `serialize_str` of a hex value which was being emitted as '0x1234'
instead of 0x1234.

Ideally, I would like to have a way to perform serialize_hex directly instead of
having to use the serialize_str function. However, I haven't been able to figure
out how to implement a custom serialize function which is not supported by any
serde trait. So, pushing this change as a workaround for now.